Send and receive data from api; use modules
