--- src/modules/game.js
@@ -1,8 +1,0 @@
-class Game {
-  constructor(name, score) {
-    this.name = name;
-    this.score = score;
-  }
-}
-
-export default Game;
